move extensions to a module
