+++ src/descriptor/bare.rs
@@ -32,7 +32,7 @@ use {
     TranslatePk,
 };
 
-use miniscript::ext::NoExt;
+use NoExt;
 
 use super::{
     checksum::{desc_checksum, verify_checksum},
