+++ src/descriptor/segwitv0.rs
@@ -30,7 +30,7 @@ use {
     TranslatePk,
 };
 
-use miniscript::ext::NoExt;
+use NoExt;
 
 use super::{
     checksum::{desc_checksum, verify_checksum},
